authorization implemented

+++ src/api/notes/handler.js
@@ -16,7 +16,8 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload;
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId });
 
       const response = h.response({
         status: 'success',
@@ -48,8 +49,9 @@ class NotesHandler {
 
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -61,6 +63,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+    
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -92,6 +97,8 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId} = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
  
       await this._service.editNoteById(id, request.payload);
  
@@ -122,13 +129,15 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
       return {
         status: 'success',
         message: 'Catatan berhasil dihapus',
       };
     } catch (error) {
-
+      //console.log(error)
       if (error instanceof ClientError) {
         const response = h.response({
           status: 'fail',
